Web/Backend/requirements.txt
pip==20.2.2
Flask==1.1.2
# gunicorn==20.0.4
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
blinker==1.4
certifi==2020.6.20
click==7.1.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
flash==1.0.3

Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8

isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.1
numpy==1.19.0
pandas==1.0.5
pipenv==2020.6.2
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
schedule==0.6.0
simplejson==3.17.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
virtualenv==20.0.26
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1